In progress.
The arc serializes to a correct-looking recipe, but it doesn't typecheck.
This seems to be because the immediate handle in the original recipe
creates an in-arc handle that includes a shape with an unresolved type
variable reference.

That's a bug. However, it's also wrong for this to be a type variable
at all, as the immediate input should collapse the variable to a
concrete type. This is probably related to #756.