Review comments and clean-ups
Our type system used to take the following set of types: [out Product], [in Thing], [in ~a] and assign ~a to either Thing or Product depending on the order in which the types were provided.

This in turn was causing recipes to mysteriously fail to pass through certain stages of the planner, as type ordering dependence generated strategy ordering dependence.

This rewrite uses upper and lower bound constraints instead of assigning concrete types to type variables. In the case above, ~a would be "at least" Thing and "at most" Product as a result of this resolution.