add extra debugging, extend timeout to improve stability (#1178)
I believe that 02edcae contained a regression. As part of that cl I added code to have the waitForVisibible() method check for both selectability & visibility (previously it had been detecting selectability).

Waiting for the element to be visible seems to take a bit longer (so this increases the wait), and adds another source of errors (so this adds more logging for future diagnosing).

The code I added in 02edcae also may have had a bug - instead of a boolean I was returning a Promise from the closure into browser.waitUntil(). It's not clear if browser.waitUntil() was resolving the Promise or using that as a truthy value. Now, however, this avoids async code entirely.