give appveyor builds a clean way to run all non-web tests (#1206)
Given the flakiness we've seen (currently tracked in #1189), I'm not
sure that we're anywhere close to being ready to run the selenium tests
in more than a single autobuild, nor am I sure there's any benefit to
doing so. We can revisit later.

Instead, this puts in place a cleaner system that appveyor can use to
run only the non-web tests. This will fix #985.