Reland: Make manifest ids unique via a session id.
Ensures that, for example, store ids for stores with the same name
loaded from a manifest file with the same name receive differing ids,
as they should, since their values could differ.

Updated for reland: Added logic and unit test to ensure that recipe
digest stays the same across multiple parses regardless of immediate
or entity handle manifest session id (as this is important, for
example, for replay purposes).