try rolling back to npm 5.7.1 to address #1242 (#1261)
This was our previous known-good version (any further back than this had other issues that were fixed in #986).

Unfortunately, it's not clear that this will make the build more stable, so I'm not sure this is worthwhile. It's something, though. Also as part of this I've run `npm update` to bring everything to latest (there were some other new dependent versions that were automatically brought in, so I figured it was better to just pick up the rest as well).