Remove `version` from Particle.onHandle* handlers
We should avoid passing this to particles this until we have a need
for it. Once we start accepting local mutations at the storage-proxy
or storage object the notion of version gets hairy (remote vs client vs pec).

An alternative would be that we drop this PR and maintain some notion
of version. e.g. short-term the storage-proxy starts at 0 and increments
each time it sees a mutation (both internal and external).