Allow node modules to be loaded using non-filesystem paths from non-node-modules
Currently the module loader allows an import like

import {Foo} from 'Bar'

If 'Bar' lives in the node_modules directory and the file that is referencing
it is also located in the node_modules directory, otherwise an error is thrown.

This change allows node_modules to be loaded without filesystem paths from
any module.