Add restrictive policy for CSP
This is only statically embeded for now, and only in the main
index.html to support static file serving. Later PRs can change
this to some build-time step to make it more manageable.

As is, this locks down as much as possible, external resource
requests, but poses a headache for adding more remotely hosted
3P particles and services. We will need to devise a declarative
scheme to allow particles to declare what external resources they
need, as well as a pre-load strategy to prevent them from being
used as a covert channel.