Comvert recipe subdirectory to typescript (#2111)
Also removed 2 typescript lint rules:

no-angle-bracket-type-assertion:
This lint rule is only relevant if the project uses .tsx, which we don't. `<string>foo` is closer to other language casts than `foo as string`, so it's worth allowing this syntax.

no-conditional-assignment:
conditional assignment is used pretty heavily in recipe comparators and they're a lot more readable as a result. I think the benefit of keeping this technique outweighs the potential costs of accidentally performing an assignment in a conditional expression.