Make innerArc.loadRecipe(...) return ids of provided slots. (#2384)
Required to iteratively build recipes inside innerArcs on top of previously loaded ones.

(To get there fully we'll also need to return ids of instantiated handles (i.e. create handles), but that's a story for another PR.)

Short term this will unlock CollectionRenderer (#2236) to operate on mutable collections of items.

Proposal in slightly more detail:
When loading a recipe, transformation particle gets back a ids of slots provided by particles in the recipe in the format 'particleLocalName.slotConnectionName' => id. This does require the loaded recipe to put local names on particles to receive the slot IDs back - which is required to disambiguate between multiple instances of the same particle in the most particle-developer friendly way.