Make the manifest parser output more consistent for primitive schema types.
Currently the parser outputs a plain string as the 'type' node for primitive fields (Text, Number, etc.) and an object with kind, location and other fields for collections, unions, etc. This complicates the processing of schema types and will make dynamic proto generation harder as we add support for the more complex types.

This CL makes the parser produce objects like {kind: 'schema-primitive', type: 'Text'}.