Rework the Entity interface.
Currently any fields or methods defined on Entity (like 'id' or
'identify') will clash with fields of the same name in a schema. To fix
this, we'll "hide" all such internal fields using a Symbol key, so only
schema-based fields are ever directly set on the Entity object itself.

This complicates access to the internals. Entity now provides static
accessors to them, and Particle has a small number of helpers around
those. I've left a few trap getters to try and catch incorrect uses of
id, rawData and dataClone - these will be removed fairly soon.