Added Flow, FlowCondition, FlowCheck, FlowModifier classes.
These classes are all essentially eqivalent to the
Check/CheckCondition/Claim classes we already have on the ParticleSpec
class. However, instead of pointing to complex objects like
HandleConnectionSpec, they are now just all strings (using nodeId,
edgeId, tag strings).

This will be incredibly useful for testing the new Hindley-Milner
algorithm, since we won't need to create complex Handle/Particle
objects, but can operate on simple strings instead.

It also makes some of the code simpler and more efficient (all of the
'is from handle'/'is from store' checks now get evaluated only once,
whereas before they were evaluated for every path).

Also note that the Flow object is going to be especially useful for HM.
It currently represents a path, but in HM it's going to be used to
represent the accumulated set of claims for every edge (with few/no
changes at all).