Added a DeepSet implementation
This is like a regular Set, except instead of using pointer equality to
determine whether elements are in the set, it uses value equality. Value
equality is implemented via a toUniqueString() hack, for
simplicity/efficiency. Every element in the set has a unique string
representation, and the string representation is used to determine
whether other elements are duplicates.

Also adds some new methods to Flow and FlowModifier, and adds FlowSet
and FlowModifierSet using DeepSet.