Wasm C++ unit test integration (#3315)
* Wasm C++ unit test integration

Test directories that hold a file called 'wasm.json' will cause sigh to
generate the C++ header file holding the entity classes from a given
manifest, then invoke emscripten to produce a wasm module.

The manifest is also linked into the build output dir to make it easy
to load the generated wasm module from it. This means unit tests will
need to load that manifest from the build dir as well.

This adds another npm dependency to pull in emsdk. The API for using
that is currently a bit awkward and will hopefully be improved in the
future (likely by us).

* Log output on failure to extract emsdk version

* Comment on regex used to extract emsdk version

* Update with Kotlin code gen; add outDir and linkManifest to wasm.json

* fixed regex version check

* Minor review fixes

* Workarounds to get this working on Travis

* Use 'xenial' environment on Travis