Added the concept of "ingress" into nodes. (#3351)
Currently there is no way to detect ingress automatically. It has to be
done manually in the tests themselves. In the future, ingress nodes will
be handles with the create or map fate (i.e. pointing to external data
sources), and possibly others too.

A check will now fail if there is no ingress into the edge. "No ingress"
is just the case when there's no flow (empty FlowSet).