Added support for cycles to the dataflow analyser! (#3360)
Cycles are dealth with by the new removeSelfReference() method. Self-modifiers are applied to copies of all other resolved and unresolved flows in the EdgeExpression. Dealing correctly with ingress falls out naturally from the other EdgeExpression operations; we don't have to do anything in particular for that.

Added lots and lots of tests: unit tests for the removeSelfReference method, and tests involving recipes with cycles.