Fixes bug when computing 'derives from' edges. (#3377)
The bug arises when processing a claim like 'claim foo derives from
bar'. If the edge 'bar' has not been created yet, then we can't store it
in the derivesFrom list for the edge 'foo' (and instead store
undefined).

Now, we compute the derivesFrom list in a separate pass after all edges
have been constructed.