Added an OrderedSet class. (#3385)
* Added an OrderedSet class.

Hooks it up in the Flow/FlowModifier to use for edgeIds. This will be
needed in order to resurrect the full path of edges (in order) from the
Flow, so we can report sensible failure messages.

* Reverted ordering of edgeIds in toUniqueString() methods.

This was breaking the deduping of equivalent cycle traversals, and
exploding the search space.