Tweaked particles symlink in javaharness.
particles/WasmTemplate had a BUILD file with paths that don't resolve in the javaharness WORKSPACE file. I've deleted the symlink to particles and replaced it with a folder with a PipeApps symlink inside it. If we need other folders we can add additional symlinks.

This should let us run `bazel build ...` and `bazel test ...` from inside the javaharness folder without bazel query errors.