Adds IArcsService.registerRemotePec() and RemotePec class. (#3587)
This is some basic functionality for hooking a remote pec up to the main
pec. In theory messages should be sendable to and from the remote pec,
although I haven't tested this yet.

Also rejiggered the ArcsServiceBridge async code quite a lot. It still
looks overly complicated to me, I think there's a better way of doing
it.

There's a lot of TODOs scattered around :-(