Updated RemotePec to allow starting and stopping of arcs.
We can't actually stop a running Arc (the Arcs Runtime doesn't support that), so all we can do is clean up the PEC (both the instance running in the remote service and the RemotePecPort running in the arcs service).

Also tidied up RemotePec.init() to be runArc() and to use proper IDs.