Refactor StorageProviderBase.on() and .off()
These methods now have the same API as the new storage stack's on and
off methods. The old behaviour is preserved in legacyOn and legacyOff
methods (there's a bunch of tests that rely on that old behaviour, but
they are all specific to the old storage stack and so will be deleted
later).