Notify ALDS upon app disconnection (#3761)
Currently the disconnection can be noticed by ALDS only when the app
is killed. The in-app disconnection would only unbind and destroy
the Arcs service implies the web content/context at the native engine is still
breathing there. The stale web content/context needs to be cleaned up bundling
with the service death.