ALDS should be off by default for the demo app
With default on, the demo app would wait until the Arcs Explorer
at the host connects to the ALDS since the Arcs runtime at the
device blocks waiting for the ready signal of the Arcs Explorer
at its initialization path by design of explore-proxy parameter.

The android property "debug.arcs.runtime.use_alds" has been ready
in place to switch the ALDS proxy dynamically, so we can deem
ALDS on/off as an option as of now.