Move src_kt into the main repo WORKSPACE
The sources are still all in src_kt (we should probably move them one
day). And src_kt_runtests.sh is still there, but it isn't used anymore
for running tests on travis.

Getting closer to the goal of having everything be in the root
WORKSPACE, and being able to run `bazel test ...` on Travis to test the
whole repo.