Preserves directory structure in arc_manifest_bundle (#3894)
Updated arc_manifest_bundle build rule to be a proper rule instead of a
macro. This allows us to use more advanced bazel functions to
reconstruct the expected directory structure for Android assets folders.

Also removed the generate_root_manifest.sh script, since it's no longer
needed.

Added a manifest bundle as an example to the javaharness. I'll replace
it soon with the actual manifests needed by javaharness.

For #3819