Fix android demo app, configure shell (#3935)
The shell-android API was modified in #3925 to require sending a 'configure'
message after receiving the 'ready' message. This PR adds that, and
mostly restores functionality in the demo app.

Also updates the demo app to use the generated root manifest
demo_manifest_bundle.arcs, and fixes the generation script (semicolons
caused it to break).