Adds `external` keyword to particle defn syntax
This indicates that the particle is not instantiable, and does not have an impl file. Used for Android particles.

Formerly we would say:
```
particle Foo in 'path/to/nowhere.java'
```
and it would magically work because it ends with .java (the file doesn't even have to exist).

Now we can say:
```
external particle Foo
```