schema2cpp: abandon inheritance and use a simple template ctor (#4059)
* schema2cpp: abandon inheritance and use a simple template ctor

Using inheritance to mirror the type lattice Just Doesn't Work(TM) in C++
because even completely abstract classes aren't really interfaces. So,
if two unrelated base classes have a common field, this introduces
ambiguity even if they are just declaring an accessor method and don't
have the field actually defined. Sigh.

So, no more inheritance. All entity classes are stand-alone, but with a
template constructor which allows slicable entities to be implicitly
converted at call sites, which is reasonable (and actually a lot easier
to implement).

* Removed TODO, added comment