Bump emsdk commit number
As mentioned in #3968, it looks like emsdk have changed their storage host, so the old AWS URLs no longer work. Bumping the commit number we're pinned to so that we can point at the new correct URLs. Note that I haven't changed the actual emscripten version we're using, so the compiler should still be exactly the same.