ConCap: test tutility to capture or silence console logging. (#4086)
Some tests result in console output as a normal result of the test case
(usually due to error logging). This tends to obfuscate the output of
the test execution, and sometimes hides genuine problems because we end
up ignoring them amidst the noise. The ConCap class provides a simple
mechanism for disabling console output completely, or capturing it to
verify that specific messages were logged.