Remove a bunch of internal keywords
These internal functions are called in unit tests, but that is not supported yet in google internally. I've marked them with `/* internal */` so that we can revert them later.

Using `@VisibleForTesting` would be a good compromise too, but there's a few hoops to jump through to get that working, so let's just wait until `internal` works.