Break cycle between Manifest and Synthetic Storage.
Moves cycles from 8 to 7.

Introduces a HandleRetriever class used only for synthetic storage so
as to break the direct link to Manifest. This in turn requires passing
a retriever in to the (old storage stack) StorageProviderFactory. We
could then pass it along to the individual storage drivers even though
not all of them need it, but given the stack migration underway, I
decided to just add an accessor to the factory.

Only synthetic storage even has the need to retrieve the handles from
a manifest, so there may be a better factoring, but this seems
reasonably well contained and a small change.