Rename wasm classes to include "Impl" suffix.
This is for parity with the JVM ones, e.g. SingletonImpl.

The convention is that shared interfaces have a normal name
(`Particle`), the JVM implementations will have a Impl name
(`ParticleImpl`) and the wasm implementations will have a Wasm prefix
too (`WasmParticleImpl`).

Disambiguating these classes by using different names is necessary
instead of using expect/actual keywords, because the signatures are
different. e.g. WasmSingleton requires a WasmEntitySpec instead of just
an EntitySpec, since it needs the decode method.