Add type parameters to DirectStore and BackingStore (#4464)
Two core changes:
* Add type parameters to DirectStore
* Add type parameters to BackingStore

The rest of the changes are just adapting to these types requiring
parameters now.

Practically, this is not increasing the type safety yet. But it's a step
down that path. Whether or not it's worth continuing down this path is
up for debate.

Areas specifically avoided for now:
* adding ReferenceModeStore type parameters
* type safe construction patterns