Add VersionMap proto
This proto will be used in the SQLite database to store serialized version maps as BLOBs.

I've also deleted the ParcelableVersionMap class and have replaced it with proto serialisation. This has reduced a ton of code. I'm hoping to replace the other handwritten Parcelable classes with protos too, assuming we like this new approach.