Plumb local CRDT application state back to handles.
If an operation fails to apply /locally/, the developer needs to
retrieve the latest version map (by reading the handle), and then retry
the operation. Alternatively, they were doing something invalid, like
removing an entity from a collection more than once. This also needs to
be plumbed through to the developer facing handles, but those are
currently in flux so I've held off until Ray's PR lands.

Note: if the update applies locally but NOT at the store, this is still
success from the developer point of view. The storageProxy will trigger
a sync with storage, and the resulting model will be merged, preserving
the locally generated change.