Stop throwing exceptions when DatabaseImpl encounters an old version (#4841)
* Stop throwing exceptions when DatabaseImpl encounters an old version

insertOrUpdate now returns a Boolean instead of a version number Int, which indicates whether the operation was applied or not. Previously we would throw an exception if the version number given in the request was the same or older than the one stored in the database. Now, getting an old version number is a no-op, and will return false.

* Fixes for failing tests

* lint