Manage client store callbacks directly in the StoreService
When the client registers a callback with a `ServiceStore`, forward that
callback to the StorageService for registration in the appropriate store
on the service side.

Removes the need for a single channel of callbacks between each service
store and binding context, and removes the need to manage proxy
callbacks on the client side.