Create DatabaseManifest
Creates DatabaseManifest, as well as AndroidSqliteDatabaseManifest as a way to persist a list of all databases created by arcs. Until now, the only list of databases created by arcs was kept in-memory and would not reflect databases which had been created previously but hadn't yet been bound-to in the current runtime.