Use random callback IDs
And fix one test that broke with this change (a callback from the
container store via the reference store would be suppressed for the first
callback attached to the RefModeStore, but not for subsequent ones).

There were a few bugs that would be masked in simple single-callback
tests because callback IDs all happened to line up, since they all
started at 1.