recipe2plan: `create` handle fixes (#4957)
## Summary
`recipe2plan` should only try to create storage keys for `create` handles that have IDs. If a recipe does not provide an ID for these handles, they will need to be created by the allocator. 

This PR also fixes an issue that @cromwellian pointed out: Now the `plan-generator.ts` will use the same identifier for all HandleConnections connected to the same handle.

## Changelog
* Storage Keys should be created for `create` handles with IDs.

* Storage Keys should be created for `create` handles with IDs.

* Added a test for id requirement

* Added test to ensure propert `CreatebleStorageKey` ids are created

* impl create handle requirement

* lint fix

* Update goldens

Co-authored-by: Alex Rosengarten <alxrsngrtn@google.com>