Remove FailureParticle (#4972)
Instead of creating a FailureParticle, just throw an exception.

This construction was only used in one place (instantiateParticle), and likely indicates a programmer error in Arcs itself that we should know about (and which should bring down the process), rather than an error in the particle author's code. So I've removed it and replaced it with an exception.

Some of the error-handling logic surrounding this becomes simpler as a result.