Writes a specification test for generated entities.
This test serves as a contract for the code generated Entity classes. It's purpose is to document the API of the generated classes, to document whether a given feature has been implemented (eventually, for now there is only the one test file), and to ensure that the generated code doesn't change without intention.

I'm hoping that others will add more specification tests for their own features as time goes on.

I'll figure out what's going on with collections and add those in a later PR. They don't seem to work properly.