Change `ProxyCallback`s to return no value (#5011)
There's no need for a Store to act immediately on a result from a
callback to the StorageProxy (or other callers). In fact, there's
already no usage of this return value.

If a `StorageProxy` (or other Store agent) encounters an issue applying
the callback message from a `Store`, it doesn't need to signal that
immediately, instead it can send the appropriate message back to the
`Store` (as is already happening).

This also removes `ParcelableProxyMessageChannel` (rather than also
refactoring it) since it was not being used anymore, except for in a few
tests.

If an AIDL-defined method doesn't expect any response, or provides a callback to get a response, use the oneway keyword to allow asynchronous invocation.