Deprecate StorageHandles and remove usage from SDK Handles
SDK Handles now stand on their own, without wrapping the `Handle` in
arcs.core.storage.

* EntityHandleManager is now the main entry point for using handles.
* EntityHandleManager now manages its own set of StorageProxy instances.
* SDK handles now work directly with storage proxy
* Small tweaks to `AbstractArcHost` to make it simpler to customize the
  HandleManager.
* TTL & Identity validation are now wrapped in a class called
`EntityPrep`

I think this code should be broken up into files, but I'll do that in a
subsequent PR, so as not to make reviewing even more difficult.

As excited as I was to also remove 100s of lines in this PR, we we are
not quite prepared yet:

* Allocator is using StorageHandles directly for persisting Arc info
* One internal demo uses storage handles directly (likely it can be
  deleted at this point)
* Two sets of integration tests use storage handles directly
* Removal Manager uses storage handles directly
* I punted on merging the reference implementation, so SDK Reference
still wraps a storage-layer Reference object.