Fix different-handle-manger tests, fix uncovered issues
* A bug in the helpers in `HandleManagerTestBase` meant that the
`DifferentHandleManager*` tests were not actually testing different
handle managers. Fixing this uncovered some failures.

* Singleton clear operations were not being signaled across the driver
layer because of a small issue in a reference mode store `Message`
translation helper.

* When `CrdtSet.merge` resulted in removing items in the local replica
due to a removal from the remote replica, the item was removed, but the
corresponding was not reported in the `MergeChanges.myChanges` result
set.