Fix broken e2e singleton test.
Re-enables the broken test (reverts #5161).

Serialisation bug was uncovered by #5145, which converted part of the serialisation code to protos. The bug was that the CrdtSingleton.Data interface implements CrdtSet.Data, but the when-statement was not ordered accordingly. So when trying to serialise the data, it would misinterpret it as CrdtSet.Data.

The real fix should be to break that inheritance chain, but I couldn't find an easy way to do that, so this will have to do for now. I modified some unit tests to improve coverage of the buggy code.

Also fixed some incorrect toString debug messages, and put the test app's UI inside a scroll view, so that you can still push the buttons when the debug text is long.