Stop extending CrdtSet.Data from CrdtSingleton.Data
This makes the fix introduced in #5167 no longer necessary, since CrdtSet.Data and CrdtSingleton.Data are now completely distinct types.

A lot of code relied on CrdtSingleton.Data being able to be cast to CrdtSet.Data, so I introduced an asCrdtSetData() method to make that easy to do.

Includes a small refactor removing the dataBuilder, which never seemed to be used for anything.