Fix arcs_kt_gen duplicate symbols issue
When building the "combo library" rule, the jar was dpeending on the
previously-generated "<name>_schema" library that had just been built.

When building Android projects, this triggered a "duplicate symbol"
error, since the symbols from the included library, along with the
re-generated symbols for the combo library were being included in the
final jar for the combo library.