Bazel update to arcs_kt_gen, arcs_manifest, and arcs_manifest_parse_test (#5163)
## Summary

### `arcs_manifest` is now a simple bundler (i.e. filegroup)
Manifest checking has to be decoupled from manifest bunding because the `arcs_manifest` rule cannot take in Jvm source files (causes cyclic dependency). Thus, the manifest checker functionality had to be pulled out. 

### `arcs_manifest_parse_test` will verify Jvm partiles
This improvement involved changes to the loader to verify Java-style classpaths. In addition, the `manifest.ts` was modified to refer to the abstract base loader (more generic). Lastly, I updated the manifest checker CLI to print help docs on `--help` or no-arg invocation.

I added a few examples of use of this rule in the android demo and in PipeApps. Now, manifests should be passed in to `srcs` and particle code should be passed into `deps`. 


### `arcs_kt_gen` supports manifest dependencies
Developers may want to write `Schema.arcs` files that contain shared schemas. This breaks Ben's `arcs_kt_gen` tool, which also offers a lot of developer convenience. 

This change allows one to specify an .arcs file as a dependency in `arcs_kt_gen`. 

## Changelog

* Modifying arcs_kt_gen to allow passing schemas as dependencies

* Added test case for codegen

* passed buildifier

* fixed manifest parse test

* Manifest checker now checks Jvm classes

* updated usage docs

Co-authored-by: Alex Rosengarten <alxrsngrtn@google.com>