Introduce ArcHost configuration, with migration path
This introduces the structures and constructors needed to present the
configuration-object based approach to ArcHost creation.

This change doesn't require that all usage points switch over to the new
method. A migration path has been provided. Fields and methods providing
backwards compatibility have been marked with deprecated.

When this is merged, new code can use this pattern, and old code can be
migrated in a piecewise fashion.