Adds support for type slicing in Entity deserialisation. (#5201)
Example of type slicing: a particle writes Person {age, height} to a handle, and another particle reads Person {age}. Previously this would throw an exception when deserialising, because height is not defined. Now unknown fields in the RawEntity will now be silently ignored when deserialising. This is a slight hack, but the behaviour is consistent with what we do for wasm.

This PR is paired with cl/308174642