Remove SchemaRegistry.getEntitySpec().
The idea was unsound and was causing errors. It assumed that there was a 1:1 mapping between schema hash and EntitySpec, which is not true. We generate different EntitySpec classes (and Entity classes) for each particle, but if they use the same schema they will have the same schema hash, and will collide.

There was only one usage, so I worked around it by giving each Entity class it's own map of hash -> EntitySpec for all of the Entities classes it references.