Fix two issues with storage proxy (#5392)
(1) duplicate onSync message when first setting up. This was happening
because we call requestSynchronization() inside registerHandle. This
can cause a message to be delivered to the proxy which causes it to
become synchronized. When we register a new handle, if the proxy is
already synchronized before the handle is added to the list then we need
to send a "fake" onSync so the particle gets the right sequence of
messages. However, if we don't check synchronization state until the
end of the method then the message delivered as a consequence of
requestSynchronization might result in there being duplicate onSyncs,
one from that message and one from this check.

(2) swallowed updates. When the store sends a model to the storageProxy,
we were only delivering it upwards if the model wasn't already synced.
However, this loses state that should be converted to an update message.
Sometimes we won't be able to deliver an update, because we can't
recover the change as a set of operations. In these cases, we still
need to deliver the udpate as a sync; so we need to desync first in
order to do so.