Utility to convert a predicate to Disjuntive Normal Form (DNF).
This PR renames `asSetOfBitSets` into `asDNF`, which was its original
intention.

However, it turns out that the implementation of `asSetOfBitSets` did
not provide a way to correctly represent negated literals in a
conjunction. e.g., it was not possible to represent `(A and B)`.
This PR introduces an addition BitSet, whcih acts as a mask to
identify which bits are valid.