Utility to convert a predicate to Disjunctive Normal Form (DNF). (#5406)
This PR renames `asSetOfBitSets` into `asDNF`, which was its original
intention. However, it turns out that the implementation of `asSetOfBitSets` did not provide a way to correctly represent negated literals in a conjunction. e.g., it was not possible to represent `(A and not(B))`. This PR introduces a Conjunct data class to deal with this limitation.