Extract logic related to partition serialization into a separate object.
This helps to clarify the actions that the Allocator performs, and opens
up the possibility of providing alternate partition-serialization
strategies, if desired.

This also removes the parallel "Map<ArcId, Partition>" which seemed to
be used for testing, or to provide partitions to an arc. Instead, the
partitions used to create an Arc object are now stored directly in that
Arc object.

Note that as a part of this change, `Allocator.create` is no longer a
suspend function, which will make it easier to construct these.