Remove getLocalData from ActiveStore, StorageService interface (#5510)
This functionality isn't needed, it is present because it's part of
`ActiveStore`, from which `ServiceStore` inherits.

This is a step in the process of making the StorageService interface
match that of `StorageCommunicationEndpoint`.

* Remove `getLocalData` from AIDL, and associated implementations.
* Remove `getLocalData` from `ActiveStore` interface.
* Remove the entire `ServiceStoreTest` file. It had one test that was
disabled for flakiness. Functionality of the service store is tested in
other ways.

The only non-testing usage of getLocalData is in ReferenceModeStore.
It's probably possible to remove this, but perhaps not worth the effort,
given the limited life remaining for that component.