Support importing .arcs files in manifest2proto (#5533)
Also introduces `Manifest.validateUniqueDefinitions()` to verify that there are no duplicates in the imported files. It'd be nice to have proper namespaces one day, so I don't want to go too far in that direction right now, so for now it's just a method on `Manifest` that you can optionally call (and we do call it, for manifest2proto).