Deflake Core SameHandleManagerTest (#5535)
* Remove Ignores from Core's SameHandleManagerTest.
* Add monitor handle manager, use that to wait for entities to make it to ramdisk.
* Fix merge issue with CrdtSet. Use new Store for each dereference.

An issue existed where a CrdtSet<RawEntity> wouldn't update the entity value for a set entry with the same id, but a newer version.

Another issue existed where the cached/stashed store from the StoreManager in the RawEntityDereferencer wasn't updated with the latest value from ramdisk yet. It makes more sense anyway -for now - to create a new Store with each dereference call.