Deflake Core DifferentHandleManagerTest (#5544)
* Remove the ignores.

Dont crash if the containerStateChannel is closed when we're trying to close it. Use best practices for collection_removingFromA_isRemovedFromB

QueryCollectionHandle's query method should not be suspending.

Be sure to send an onUpdate after a resync when the consumerView changed.

It's okay to try and get the state while we're de-synced.

I think we're missing ops that are sent after the resync modelupdate is sent but before we've processed it.

Fix the ordering issue in the StorageProxy. Add Tests for it.

Filter the model ops to apply.

Use a different mechanism to determine whether or not we should send an update event.

Fix some practices. and fix some access of the local data from the RMS's container store.

* Be more careful about how we shutdown the schedulers.

* Log at debug when we try to close the closed channel.

* Use a ConcurrentHashMap instead of requiring locking for the JvmScheduler map management.

* Resolve PR comments.

* Opt to not use toType from Handle.kt, instead - create the type directly.

* Whoops.