Tweaks to fix particle lifecycle init race conditions (#5593)
When initiating a collection of Particles, it's possible that particles
will contain handles that share an underlying storage proxy. Currently,
a particle moves to the Running state when all of its handles move to
the ready state.

There are two issues addressed that are related to the fact that a
particle may be initialized before `runParticle` is called for it:

1. notifyReady will not have been set on the particle yet, which means
that despite the fact that the particle has reached the ready state, the
notifyReady method that gates completion of performParticleStartup will
never execute. This is addressed by calling the method immediately when
it's set, if the particle is already in the Running state.

2. Because runParticle checks for a particle to be in the "Waiting"
state, but it may have already reached the "Running" state, triggering a
crash when the state is checked.