Remove the `Store` wrapping type.
It is a pattern that we haven't used effectively on the Kotlin side.

The bulk of this PR is removing the `Store` class completely, and
changing:
`Store(options).activate(activationFactory)`

into

`activationFactory(options)`

The StorageService was taking advantage of the deferred nature of
`activate` to prevent constructing Stores on the main service thread;
this is handled in this PR by introducing a simple deferred store
constructor class for that purpose.