first commit dsm-callback
modified ProxyCallbackManager/DirectStore to allow the observer to generate the callbackToken

laying foundation to allow a set of communication channels per store in DirectStoreMuxer

removed onWithToken

modified proxyMessage to include nullable muxId

added an exception to ProxyCallbackManager register method

created proxyManager for DirectStoreMuxer and enforced observers (atm just the ReferenceModeStore) to register to the DirectStoreMuxer

Attempted to fix some race condition bugs. Part of that is changing getLocalData. And I aslo made muxId immutable.

The referenceModeStore relies on the proxyMessage retaining the same proxyMessage Id when being sent to the containerStore (rather than overriding the id with the containerStore's id). This is weird but we're getting rid of reference mode store anyways soon so I don't think it matters?